fix: 훅 분리 #40
